Change basic NIC to use the new infra_basic library (copy of infra_vr)
Do some mods to infra_basic: sending pkts to the wire and handling host flags

Also change test code to create bigger pkts to more fully test csum code

+++ firmware/lib/nic_basic/_c/nic_switch.c
@@ -85,7 +85,7 @@ nic_switch_reconfig()
 
 
 __intrinsic int
-nic_switch_rx_defaultq(int vport, uint8_t *qid)
+nic_switch_rx_defaultq(int vport, uint32_t *qid)
 {
     __shared __lmem volatile struct nic_local_state *nic = &nic_lstate;
     int ret = NIC_RX_DROP;
